fix(core): align Gesture action contract
