+++ packages/core/src/core/ui/gesture/core.ts
@@ -1,18 +1,23 @@
-import type { InputAction } from '../input-action/input-action';
-
 export type GestureType = 'tap' | 'doubletap';
 export type GesturePointerType = 'mouse' | 'touch' | 'pen';
 export type GestureRegion = 'left' | 'center' | 'right';
+export type GestureActionName =
+  | 'togglePaused'
+  | 'toggleMuted'
+  | 'toggleFullscreen'
+  | 'toggleSubtitles'
+  | 'togglePictureInPicture'
+  | 'toggleControls'
+  | 'seekStep'
+  | 'volumeStep'
+  | 'speedUp'
+  | 'speedDown';
 
 export interface GestureProps {
   /** Gesture to recognize. */
   type: GestureType;
-  /**
-   * Player action to run when the gesture is recognized. Built-in actions are `togglePaused`, `toggleMuted`,
-   * `toggleFullscreen`, `toggleSubtitles`, `togglePictureInPicture`, `toggleControls`, `seekStep`, `volumeStep`,
-   * `speedUp`, and `speedDown`.
-   */
-  action: InputAction;
+  /** Built-in player action or same-named custom store action to run when the gesture is recognized. */
+  action: GestureActionName | (string & {});
   /** Numeric value passed to actions such as `seekStep` and `volumeStep`. */
   value?: number | undefined;
   /** Pointer type that may activate the gesture. All pointer types are accepted when omitted. */
