+++ packages/core/src/core/ui/tests/gesture-actions.test-d.tsx
@@ -1,0 +1,14 @@
+import type { GestureProps } from '../gesture/core';
+
+const builtInGesture = {
+  type: 'tap',
+  action: 'toggleControls',
+} satisfies GestureProps;
+
+const customGesture = {
+  type: 'doubletap',
+  action: 'openTranscript',
+} satisfies GestureProps;
+
+void builtInGesture;
+void customGesture;
